Order Command | Error Logic
Added the order command. User will run the command, and as long as they don't have an order open in the database an order channel will be opened up. This channel is only viewable by staff and the user. The user may then order whatever they want.

Add error logic. In the manifest, each error code has a name, description, and level. I'm going to add a command that allows the user to query the code to attempt to solve the issue.